fix(me): name the CLI sessions page "Devices"
"Sessions" in the personal nav reads as coding sessions, which is a
different surface. The page lists machines where the CLI is signed in, so
it is now Devices: nav label, heading, page title and the in-page copy
that names what is listed.

The route moves to /me/devices too, freeing /me/sessions for an actual
sessions page later. The old path forwards to the new one so docs links
and bookmarks keep working, via a page that renders <Navigate> rather
than a `loader` redirect: loaders do not run on a cold load of the SPA,
which is exactly how a stale link arrives.
