+++ docs/ai-governance/members-and-invites.mdx
@@ -106,7 +106,7 @@ Org-level SSO configuration (IdP metadata, SAML cert, SCIM provisioning) is **se
 
 ## Where to next
 
-- **Self-service for me**: [/settings/authentication](#per-user-sign-in-methods-at-settingsauthentication) (this page) and [Sessions](/ai-governance/personal-portal/sessions).
+- **Self-service for me**: [/settings/authentication](#per-user-sign-in-methods-at-settingsauthentication) (this page) and [Devices](/ai-governance/personal-portal/sessions).
 - **Org SSO + SCIM auto-provisioning**: `/settings/governance/scim-provisioning` (admin-only).
 - **Role catalog**: [Roles & permissions](/ai-governance/roles-and-permissions).
 - **Audit who did what when**: [Audit log](/ai-governance/audit-log).
