+++ langwatch/src/components/me/InstallCliCard.tsx
@@ -22,7 +22,7 @@ import { docsUrl } from "~/utils/docsUrl";
  * snippet on cloud, and a `LANGWATCH_ENDPOINT=<host> npm install -g …`
  * variant on self-hosted (auto-detects via `usePublicEnv().BASE_HOST`).
  *
- * Drop into any zero-content surface — /me empty-state, /me/sessions
+ * Drop into any zero-content surface — /me empty-state, /me/devices
  * empty-state, governance setup checklist — instead of a bare "Run
  * `langwatch login` in your terminal" sentence that assumes the user
  * already has the CLI installed.
