+++ langwatch/src/routes.tsx
@@ -292,6 +292,14 @@ const routes: RouteObject[] = [
         ...page(() => import("./pages/me/configure")),
       },
       {
+        path: "/me/devices",
+        ...page(() => import("./pages/me/devices")),
+      },
+      {
+        // Pre-rename path for the devices inventory, kept resolving so old
+        // links do not dead-end. A page that renders <Navigate>, not a
+        // `loader` redirect: loaders do not run on a cold load of the SPA,
+        // which is exactly how a stale link arrives.
         path: "/me/sessions",
         ...page(() => import("./pages/me/sessions")),
       },
